занимались тестированием функции
