+++ composer.json
@@ -1,0 +1,8 @@
+{
+    "require-dev": {
+        "phpunit/phpunit": "^9.5"
+    },
+    "scripts": {
+      "test": "./vendor/bin/phpunit tests --colors --bootstrap=init.php"
+    }
+}
