+++ index.php
@@ -1,6 +1,6 @@
 <?php
-require_once('helpers.php');
-require_once ('data.php');
+require_once __DIR__ . 'init.php';
+require_once __DIR__ . 'data.php';
 
 $page_content = include_template('main.php',['lots' => $lots,'categories' => $categories]);
 
@@ -13,7 +13,3 @@ $layout_content = include_template('layout.php',[
 ]);
 
 print($layout_content);
-
-?>
-
-
