+++ init.php
@@ -1,0 +1,5 @@
+<?php
+require_once __DIR__ . '/functions/calculate.php';
+require_once __DIR__ . '/functions/db.php';
+require_once __DIR__ . '/functions/template.php';
+require_once __DIR__ . '/functions/validate.php';
