+++ tests/Unit/ExampleTest.php
@@ -1,0 +1,12 @@
+<?php
+namespace Tests\Unit;
+
+use PHPUnit\Framework\TestCase;
+
+class ExampleTest extends TestCase
+{
+    public function testPushAndPop()
+    {
+        $this->assertEquals(123, 123);
+    }
+}
